+++ NEWS.md
@@ -11,8 +11,9 @@
 
 - `fb_doi_from_bibtex()`: imports a BibTeX and returns DOI
 - `fb_clean_doi()`: cleans DOI
-- `fp_check_mailto()`: checks that user sends it email to OpenAlex
+- `fp_check_mailto()`: checks that user sends its email to OpenAlex
 - `fp_article_fairness()`: gets the fairness status of an article
+- `fp_journal_fairness()`: gets the fairness status of a journal
 
 ## Fix bugs